+++ dx_router/CHANGELOG.md
@@ -8,4 +8,4 @@
 
 ## 1.0.0+2
 
-- Updated example based on [dx_generator 2.0.0]('https://pub.dev/packages/dx_generator')
+- Updated example based on [dx_generator]('https://pub.dev/packages/dx_generator') (v2.0.0)